fix: move imgDet to pages
