+++ src/pages/imageDetail/ImageDetail.module.scss
@@ -1,33 +1,22 @@
-.popup {
-  position: fixed;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  top: 0;
-  left: 0;
-  overflow-x: hidden;
-  overflow-y: auto;
-  transition: all 0.3s ease 0s;
-  z-index: 100;
-}
-
 .popupContent {
   position: relative;
   background-color: #fff;
-  padding: 30px;
+  padding: 25px;
   max-width: calc(100% - 200px);
   width: 100%;
   border-radius: 10px;
 }
 
 .popupBody {
+  position: relative;
   min-height: 100%;
   max-width: 1600px;
   margin: 0 auto;
   display: flex;
   align-items: center;
   justify-content: center;
-  padding: 10px;
+  padding-top: 40px;
+  padding-bottom: 10px;
   flex-direction: column;
 }
 
@@ -64,15 +53,28 @@
   margin-bottom: 15px;
 }
 
-.closeBtn {
+.back {
+  color: inherit;
   position: absolute;
-  width: 30px;
-  height: 30px;
-  top: 4px;
-  right: 4px;
-  color: #000000;
-  background: url(../../assets/svg/close.svg);
-  cursor: pointer;
+  top: 10px;
+  left: 0px;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  transition: .3s;
+}
+
+.back:hover {
+  text-decoration: underline;
+}
+
+.backImg {
+  background-size: 20px 20px;
+  background-repeat: no-repeat;
+  margin-right: 5px;
+  width: 20px;
+  height: 20px;
+  background-image: url(../../assets/svg/arrow-left.svg);
 }
 
 @media(max-width: 1200px) {
